+++ packages/core/manifest/src/seed/services/seeder.service.ts
@@ -17,7 +17,7 @@ import { RelationshipService } from '../../entity/services/relationship.service'
 import { faker } from '@faker-js/faker'
 import * as fs from 'fs'
 import * as path from 'path'
-import bcrypt from 'bcrypt'
+import bcrypt from 'bcryptjs'
 
 import {
   ADMIN_ENTITY_MANIFEST,